价格 改变

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/sku/ProductSkuSupplyUnitRespVo.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -27,10 +28,10 @@ public class ProductSkuSupplyUnitRespVo extends CommonBean {
     private String supplyUnitName;
 
     @ApiModelProperty("无税进价")
-    private Long noTaxPurchasePrice;
+    private BigDecimal noTaxPurchasePrice;
 
     @ApiModelProperty("含税进价")
-    private Long taxIncludedPrice;
+    private BigDecimal taxIncludedPrice;
 
     @ApiModelProperty("联营扣率")
     private Long jointFranchiseRate;
